Forgotten Changes
Last commit included the addition of node modules and Ramda.js as a
dependency. This commit changes the functions placed in var(s) to
const(s) as well as a chapter 4 of Mostly Adequate file.

A few functions in chapter 3 were not abstract enough and improperly
executed. So function names and implementations have be changed.